pages_generator/requirements.txt
PyYAML==5.1.2
humanize
requests
